Move model related code into core
