+++ src/roman_datamodels/core/_metadata.py
