+++ src/roman_datamodels/core/adaptors/_adaptor_tags.py
