+++ src/roman_datamodels/core/adaptors/_base.py
