+++ src/roman_datamodels/core/adaptors/_ndarray.py
