--- src/roman_datamodels/datamodels/__init__.py
@@ -1,9 +1,0 @@
-# from ._core import *  # noqa: F403
-# from ._datamodels import *  # noqa: F403
-
-# # rename rdm_open to open to match the current roman_datamodels API
-# from ._utils import rdm_open as open  # noqa: F403, F401
-from .datamodel import RomanDataModel
-from .extended_models import RomanExtendedDataModel
-
-__all__ = ["RomanDataModel", "RomanExtendedDataModel"]
