+++ src/roman_datamodels/generator/_schema.py
@@ -7,7 +7,7 @@ from typing import Any
 
 from datamodel_code_generator.parser.jsonschema import JsonSchemaObject
 
-from roman_datamodels.datamodels import RomanExtendedDataModel
+from roman_datamodels.core import RomanExtendedDataModel
 
 from ._utils import get_manifest_maps
