--- src/roman_datamodels/pydantic/__init__.py
@@ -1,3 +1,0 @@
-from .datamodel import BaseRomanDataModel
-
-__all__ = ["BaseRomanDataModel"]
