+++ tests/test_archive.py
@@ -3,9 +3,8 @@ Run tests on the archive metadata retrieval for datamodels
 """
 import pytest
 
+from roman_datamodels.core import Archive, BaseRomanDataModel
 from roman_datamodels.datamodels import _generated
-from roman_datamodels.pydantic import BaseRomanDataModel
-from roman_datamodels.pydantic.metadata import Archive
 
 models = [getattr(_generated, name) for name in _generated.__all__ if issubclass(getattr(_generated, name), BaseRomanDataModel)]
